Move SessionButton to separate component
